IMMENSELY Optimized the Game Engine
- Cleaned up & improved a lot of code
- The game no longer lags horribly!! YEEHAW
  * things[] now only has an index for each thing, and not for each level
    unit (including empty ones), which reduces its size SIGNIFICANTLY.
      * This means that things.size() is NO LONGER EQUAL TO the number of
        level units (which used to be a key assumption of the engine).
  * Things now have tgThingsUnit (great name, I know), which
    contains which index the thing holds in the things[] array.
    tgLevelUnit is no longer valid for use when indexing things[].
- Completely fixed the camera so that it now properly scrolls when
  it should and stops scrolling when it shouldn't.
- NOTE:  THE LEVEL EDITOR IS NOW BUGGY AND CURRENTLY NEEDS TO BE
  FIXED.
- I probably did a couple other things that I forgot to log,
  unfortunately.  But maybe not.

+++ Include.cpp
@@ -72,7 +72,7 @@ bool init(void)
 	}
 	*/
 
-	for (int i = 0; i < (int)Game::things.size(); i++)
+	for (int i = 0; i < Game::things.size(); i++)
 	{
 		Game::things[i] = NULL;
 		Game::gColliding[i] = NULL;
@@ -100,7 +100,7 @@ bool initFont(void)
 
 void close(void)
 {
-	for (int i = 0; i < (int)Game::things.size(); i++)
+	for (int i = 0; i < Game::things.size(); i++)
 	{
 		if (Game::things[i] != NULL)
 		{
